fix(core): do not discard operation outputs when shutdown is set (#219)
* fix(core): operation output typing

Update typing to reflect that None can be returned by an operation

* fix(core): save operation output when shutdown set

The code assumed that if shutdown was set when an operation returned there would be no output, but this may not be the case #202

Also the logs assumed that shutdown could only be set by SIGTERM. However because of #200 we cannot for sure if this is the case. Hence the logs are temporarily changed here to reflect this.