refactor(core): separate cleanup logic from signal handling and fix nested-operation shutdown (#281)
* refactor(core): operation cleanup

Operation cleanup logic was mixed together (a single function for cleaning up everything was required/was different for different scenarios) and also mixed with signal handling (a shutdown could be successful operation or a signal).

In this commit
- cleanup code required by different operations/functions is separated
- A single signal handler exists that is generic
   - operations register their cleanup requirements with the handler so they are cleaned-up
- shutdown global now only indicates if a signal was received NOT if an operation finished

* chore(core): fix print formatting

* feat(core): Raise exception when KeyboardInterrupt caught during operation

It is raised as a new InterruptedOperationError which contains the operation identifier and is subclass of KeyboardInterrupt

Previously nothing was raised and the operation exited normally after interruption. However, this pattern is not easy to maintain with multiple nested operations and each operation would have to check if the inner operation exited due to KeyboardInterrupt.

This way operators do not have to handle KeyboardInterrupt. Each outer operation will catch the inner interrupt and raise a new exception with its own id.

The outermost handler (in operation/create.py) now catches InterruptedOperationError and prints the id of the outermost (parent) interrupted operation.

* Apply suggestions from code review

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>

* chore(core): apply suggestions from code review

* fix(core): Setting shutdown signal to true

* fix(core): Move initialize_ray_resource_cleaner

Pattern is orchestrate() creates/destroys the resource cleaner which exists in default namespace.

If another function calls an explore operation that function has same responsibility.

* refactor(core): explore operation owns resource cleaner

This is so the actors registered with the cleaner during the operation can be cleaned BEFORE the operation clean up deletes their parents (causing them to be deleted).

The other option is to have a single resource cleaner but then all actors wanting to use it must be created "detached"

* refactor(core): MeasurementQueue init

- Remove get_measurement_queue class method. It does nothing anymore
- Add ray_namespace parameter

ray_namespace parameter allows less parallel passing of same information, which potentially is source of errors, and more consistency checks.

A set of actuators, discovery space manager, and queue instance  should all be in same ray namespace.

* refactor(core): Move actuator initialization code

- Move all actuator validation and initialization code to setup_actuator.
- Create actuators first so validation failures are detected early
- setup_actuators obtains namespace from queue rather than having potential inconsistency.

* refactor(vllm_performance): Get resource cleaner via expected namespace

* fix(core): missing parameter

* Apply suggestions from code review

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>

* fix(core): shutdown operation on exception

Normal shutdown and SIGTERM were handled but other exception were not leading to operation resource being cleaned up

* chore(core): fix log levels

* chore(core): Additional logging for debug

* chore(core): Additional logging for debug

* Apply suggestions from code review

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>

* chore(ado_ray_tune): black

* chore(ado_ray_tune): code review suggestions

---------

Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>